+++ tests/Tool/Command/FileDownloadCommandTest.php
@@ -8,8 +8,8 @@ use Zalas\Toolbox\Tool\Command\FileDownloadCommand;
 
 class FileDownloadCommandTest extends TestCase
 {
-    const URL = 'https://example.com/file';
-    const FILE = '/usr/local/bin/file.txt';
+    private const URL = 'https://example.com/file';
+    private const FILE = '/usr/local/bin/file.txt';
 
     /**
      * @var FileDownloadCommand
@@ -18,10 +18,7 @@ class FileDownloadCommandTest extends TestCase
 
     protected function setUp()
     {
-        $this->command = FileDownloadCommand::import([
-            'url' => self::URL,
-            'file' => self::FILE,
-        ]);
+        $this->command = new FileDownloadCommand(self::URL, self::FILE);
     }
 
     public function test_it_is_a_command()
@@ -32,28 +29,5 @@ class FileDownloadCommandTest extends TestCase
     public function test_it_generates_the_installation_command()
     {
         $this->assertRegExp(\sprintf('#curl .*? %s -o %s#', self::URL, self::FILE), (string) $this->command);
-    }
-
-    /**
-     * @dataProvider provideRequiredProperties
-     */
-    public function test_it_complains_if_any_of_required_properties_is_missing(string $property)
-    {
-        $this->expectException(\InvalidArgumentException::class);
-
-        $properties = [
-            'url' => self::URL,
-            'file' => self::FILE,
-        ];
-
-        unset($properties[$property]);
-
-        FileDownloadCommand::import($properties);
-    }
-
-    public function provideRequiredProperties(): \Generator
-    {
-        yield ['url'];
-        yield ['file'];
     }
 }
